Cleanup and Removal of cast's

+++ Puara/ButtonAvnd.cpp
@@ -1,5 +1,4 @@
 #include "ButtonAvnd.hpp"
-#include <puara/utils.h>
 
 namespace puara_gestures::objects
 {
@@ -11,21 +10,23 @@ void ButtonAvnd::operator()()
 
   // 2. Updating configurable parameters of the underlying puara_gestures::Button object
 
-  unsigned int desired_threshold = static_cast<unsigned int>(inputs.threshold_param.value);
-  if (this->impl.threshold != desired_threshold) {
+  unsigned int desired_threshold = inputs.threshold_param.value;
+  if(this->impl.threshold != desired_threshold)
+  {
     this->impl.threshold = desired_threshold;
   }
 
-  unsigned int desired_count_interval = static_cast<unsigned int>(inputs.count_interval_param.value);
-  if (this->impl.countInterval != desired_count_interval) {
+  unsigned int desired_count_interval = inputs.count_interval_param.value;
+  if(this->impl.countInterval != desired_count_interval)
+  {
     this->impl.countInterval = desired_count_interval;
   }
 
-  unsigned int desired_hold_interval = static_cast<unsigned int>(inputs.hold_interval_param.value);
-  if (this->impl.holdInterval != desired_hold_interval) {
+  unsigned int desired_hold_interval = inputs.hold_interval_param.value;
+  if(this->impl.holdInterval != desired_hold_interval)
+  {
     this->impl.holdInterval = desired_hold_interval;
   }
-
 
   // 3. Calling the update method of the puara_gestures::Button implementation.
   this->impl.update(current_button_state);
@@ -37,7 +38,7 @@ void ButtonAvnd::operator()()
   outputs.triple_tap_event_output.value = (this->impl.tripleTap == 1);
 
   outputs.is_held_output.value = this->impl.hold;
-  outputs.press_duration_output.value = static_cast<float>(this->impl.pressTime);
+  outputs.press_duration_output.value = this->impl.pressTime;
 }
 
 }
